api/pyproject.toml
[project]
name = "derailed-api"
version = "0.0.0"
description = ""
authors = [
    {name = "VincentRPS", email = "vincentbusiness55@gmail.com"},
]
dependencies = [
    "fastapi>=0.104.1",
    "argon2-cffi>=23.1.0",
    "itsdangerous>=2.1.2",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "UNLICENSED"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
